+++ data/modules/match.lua
@@ -262,11 +262,56 @@ local function match_loop(context, dispatcher, tick, state, messages)
   end
 end
 
+--[[
+Called when the server begins a graceful shutdown process. Will not be called if graceful shutdown is disabled.
+
+Context represents information about the match and server, for information purposes. Format:
+{
+  env = {}, -- key-value data set in the runtime.env server configuration.
+  executionMode = "Match",
+  match_id = "client-friendly match ID, can be shared with clients and used in match join operations",
+  match_node = "name of the Nakama node hosting this match",
+  match_label = "the label string returned from match_init",
+  match_tick_rate = 1 -- the tick rate returned by match_init
+}
+
+Dispatcher exposes useful functions to the match. Format:
+{
+  broadcast_message = function(op_code, data, presences, sender),
+    -- numeric message op code
+    -- a data payload string, or nil
+    -- list of presences (a subset of match participants) to use as message targets, or nil to send to the whole match
+    -- a presence to tag on the message as the 'sender', or nil
+  match_kick = function(presences)
+    -- a list of presences to remove from the match
+  match_label_update = function(label)
+    -- a new label to set for the match
+}
+
+Tick is the current match tick number, starts at 0 and increments after every match_loop call. Does not increment with
+calls to match_join_attempt, match_join, or match_leave.
+
+State is the current in-memory match state, may be any Lua term except nil.
+
+Grace Seconds is the number of seconds remaining until the server will shut down.
+
+Expected return these values (all required) in order:
+1. An (optionally) updated state. May be any non-nil Lua term, or nil to end the match.
+--]]
+local function match_terminate(context, dispatcher, tick, state, grace_seconds)
+  if state.debug then
+    print("match " .. context.match_id .. " tick " .. tick)
+    print("match " .. context.match_id .. " grace_seconds " .. grace_seconds)
+  end
+  return state
+end
+
 -- Match modules must return a table with these functions defined. All functions are required.
 return {
   match_init = match_init,
   match_join_attempt = match_join_attempt,
   match_join = match_join,
   match_leave = match_leave,
-  match_loop = match_loop
+  match_loop = match_loop,
+  match_terminate = match_terminate
 }