+++ CHANGELOG.md
@@ -11,9 +11,11 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Lua runtime AES-256 functions.
 - Lua runtime token generator function now returns a second value representing the token's expiry.
 - Add local cache for in-memory storage to the Lua runtime.
+- Graceful server shutdown and match termination.
 
 ### Changed
 - Improved Postgres compatibility on TIMESTAMPTZ types.
+- Graceful shutdown is now enabled by default.
 
 ### Fixed
 - Correctly merge new friend records when importing from Facebook.